Release

+++ erd/sql/pfm-investment-related.sql
@@ -15,63 +15,63 @@ CREATE TABLE AllocationMapping
 
 CREATE TABLE InvestmentsAndSavingsDay
 (
-  id          INT          NOT NULL AUTO_INCREMENT,
-  allocId     INT          NOT NULL,
-  date        DATE         NOT NULL,
-  valueAdded  INT          NOT NULL,
-  marketValue INT          NOT NULL,
-  dateAdded   DATE         NOT NULL,
-  addedBy     VARCHAR(255) NOT NULL,
-  updateDate  DATE         NULL    ,
-  updateBy    VARCHAR(255) NULL    ,
+  id          INT            NOT NULL AUTO_INCREMENT,
+  allocId     INT            NOT NULL,
+  date        DATE           NOT NULL,
+  valueAdded  DECIMAL(15, 2) NOT NULL,
+  marketValue DECIMAL(15, 2) NOT NULL,
+  dateAdded   DATE           NOT NULL,
+  addedBy     VARCHAR(255)   NOT NULL,
+  updateDate  DATE           NULL    ,
+  updateBy    VARCHAR(255)   NULL    ,
   PRIMARY KEY (id)
 );
 
 CREATE TABLE MonthlyGrowth
 (
-  id                INT          NOT NULL AUTO_INCREMENT,
-  allocId           INT          NOT NULL,
-  month             VARCHAR(255) NOT NULL,
-  year              INT          NOT NULL,
-  contribution      INT          NOT NULL,
-  totalContribution INT          NOT NULL,
-  currentValue      INT          NOT NULL,
-  growthRate        INT          NOT NULL,
-  previousContrib   INT          NOT NULL,
-  dateAdded         DATE         NOT NULL,
-  addedBy           VARCHAR(255) NOT NULL,
-  updateDate        DATE         NULL    ,
-  updateBy          VARCHAR(255) NULL    ,
+  id                INT            NOT NULL AUTO_INCREMENT,
+  allocId           INT            NOT NULL,
+  month             VARCHAR(255)   NOT NULL,
+  year              INT            NOT NULL,
+  contribution      DECIMAL(15, 2) NOT NULL,
+  totalContribution DECIMAL(15, 2) NOT NULL,
+  currentValue      DECIMAL(15, 2) NOT NULL,
+  growthRate        DECIMAL(15, 2) NOT NULL,
+  previousContrib   DECIMAL(15, 2) NOT NULL,
+  dateAdded         DATE           NOT NULL,
+  addedBy           VARCHAR(255)   NOT NULL,
+  updateDate        DATE           NULL    ,
+  updateBy          VARCHAR(255)   NULL    ,
   PRIMARY KEY (id)
 );
 
 CREATE TABLE NetWorth
 (
-  id         INT          NOT NULL AUTO_INCREMENT,
-  allocId    INT          NOT NULL,
-  month      VARCHAR(255) NOT NULL,
-  year       INT          NOT NULL,
-  value      INT          NOT NULL,
-  dateAdded  DATE         NOT NULL,
-  addedBy    VARCHAR(255) NOT NULL,
-  updateDate DATE         NULL    ,
-  updateBy   VARCHAR(255) NULL    ,
+  id         INT            NOT NULL AUTO_INCREMENT,
+  allocId    INT            NOT NULL,
+  month      VARCHAR(255)   NOT NULL,
+  year       INT            NOT NULL,
+  value      DECIMAL(15, 2) NOT NULL,
+  dateAdded  DATE           NOT NULL,
+  addedBy    VARCHAR(255)   NOT NULL,
+  updateDate DATE           NULL    ,
+  updateBy   VARCHAR(255)   NULL    ,
   PRIMARY KEY (id)
 );
 
 CREATE TABLE YearlyGrowth
 (
-  id                  INT          NOT NULL AUTO_INCREMENT,
-  allocId             INT          NOT NULL,
-  year                INT          NOT NULL,
-  totalContribution   INT          NOT NULL,
-  averageContribution INT          NOT NULL,
-  averageCurrentValue INT          NOT NULL,
-  averageGrowthRate   INT          NOT NULL,
-  dateAdded           DATE         NOT NULL,
-  addedBy             VARCHAR(255) NOT NULL,
-  updateDate          DATE         NULL    ,
-  updateBy            VARCHAR(255) NULL    ,
+  id                  INT            NOT NULL AUTO_INCREMENT,
+  allocId             INT            NOT NULL,
+  year                INT            NOT NULL,
+  totalContribution   DECIMAL(15, 2) NOT NULL,
+  averageContribution DECIMAL(15, 2) NOT NULL,
+  averageCurrentValue DECIMAL(15, 2) NOT NULL,
+  averageGrowthRate   DECIMAL(15, 2) NOT NULL,
+  dateAdded           DATE           NOT NULL,
+  addedBy             VARCHAR(255)   NOT NULL,
+  updateDate          DATE           NULL    ,
+  updateBy            VARCHAR(255)   NULL    ,
   PRIMARY KEY (id)
 );
